Get selected index field names by their types. Connect those types to the respective condition popOver input fields. Fetch chart data only if the chart params are valid.
Assign the initial chart params only if they don't exist.

Assign the default chart params in a better way.

Fix 'over/groupped over' condition popOver menu text.

Init condition FIELD popOver based on the default falues.

Use ng-if instead of ng-show because the first selected element of the select tag "is eaten" somehow when you use ng-show in the popOver.
You see it in the dir scope, a correct query is generated but the select is empty and the element is not in the list.
And it does not appear if you select other element or close/open the popOver.

Index value in the chart_query_params is not used. Assign index value after all the index field names were fetched. It is needed
to avoid annecessary $watch() calls in the conditionPanel when there are no data about the field names.

Improve code style.

Refactor to get index fields of object type. Move the refactored fn in sentinlHelper service. Add tests for the fn.

Refactor getFields recursive fn to make it clearer.

Improve test code style.